feat: Expand and refactor AST utilities.
