+++ packages/ast/src/clone-ast.js
@@ -1,8 +1,13 @@
-import { cloneNode } from './util.js';
+import { cloneNode } from './nodes.js';
 
 export function cloneAST(ast) {
-  return {
-    metadata: { ...ast.metadata },
-    article: cloneNode(ast.article)
-  };
+  const clone = {};
+  for (const key in ast) {
+    if (key === 'article') {
+      clone[key] = cloneNode(ast[key]);
+    } else {
+      clone[key] = { ...ast[key] };
+    }
+  }
+  return clone;
 }
