+++ packages/ast/src/index.js
@@ -1,10 +1,24 @@
 export {
+  META,
+  TEXTNODE,
+  COMPONENT,
+  createNode,
   createTextNode,
+  createMetaNode,
   createComponentNode,
   isTextNode,
+  isMetaNode,
   isComponentNode,
   isCustomComponentNode,
   getNodeName,
+  cloneNode,
+  mergeTextNodes
+} from './nodes.js';
+
+export {
+  VALUE,
+  EXPRESSION,
+  EVENT,
   parseProperties,
   createProperties,
   isBacktickQuoted,
@@ -22,21 +36,42 @@ export {
   setValueProperty,
   removeProperty,
   extractProperties,
+} from './properties.js';
+
+export {
   hasClass,
   getClasses,
   removeClass,
   addClass,
+} from './classes.js';
+
+export {
   getChildren,
   appendChildren,
   prependChildren,
   removeChild,
   replaceChild,
-  cloneNode,
+} from './children.js';
+
+export {
   visitNodes,
   queryNodes,
-  mergeTextNodes,
+  queryNode
+} from './visit.js';
+
+export {
   extractText
-} from './util.js';
+} from './extract-text.js';
+
+export {
+  createRange,
+  queryRange,
+  queryPath
+} from './range.js';
+
+export {
+  setParentNodes
+} from './set-parent-nodes.js';
 
 export {
   cloneAST
@@ -45,9 +80,3 @@ export {
 export {
   transformAST
 } from './transform-ast.js';
-
-export {
-  createRange,
-  queryRange,
-  queryPath
-} from './range.js';
